Use lua-http for weather widget
If you're on questionable WiFi,
the forecast API can take a while to return a response,
which locks up the entire window manager since LuaSocket
is synchronous.  Using lua-http allows us to perform these requests
in the background - in addition, it makes the code a lot simpler.

Thanks to daurnimator for introducing me to lua-http and cqueues, and
for helping me get this written!